+++ flutter_local_notifications/lib/src/pending_notification_request.dart
@@ -1,9 +1,0 @@
-class PendingNotificationRequest {
-  final int id;
-  final String title;
-  final String body;
-  final String payload;
-
-  const PendingNotificationRequest(
-      this.id, this.title, this.body, this.payload);
-}