+++ flutter_local_notifications/lib/src/typedefs.dart
@@ -1,0 +1,8 @@
+import 'dart:async';
+
+/// Signature of callback passed to [initialize]. Callback triggered when user taps on a notification
+typedef SelectNotificationCallback = Future<dynamic> Function(String payload);
+
+// Signature of the callback that is triggered when a notification is shown whilst the app is in the foreground. Applicable to iOS versions < 10 only
+typedef DidReceiveLocalNotificationCallback = Future<dynamic> Function(
+    int id, String title, String body, String payload);